Kimchi/constraints: move into top level tests directory
More rustacean
